# api/openapi.yaml
openapi: 3.1.0
info:
  title: Worthly API
  version: 4.3.0
  description: |
    Canonical MVP API contract for Worthly.

    Protected resources are Bearer-token OAuth2 Resource Server endpoints.
    The Next.js `web_session` cookie is never an API credential.

    Aggregate amounts are always grouped by ISO 4217 currency. No schema
    exposes a single cross-currency Money total.
servers:
  - url: https://worthly.example.com/api/v1
security:
  - oauth2: [worthly.read]
...
paths:
...
  /connections/enable-banking/callback:
    get:
      tags: [Connections]
      operationId: enableBankingCallback
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
        - {name: error_description, in: query, schema: {type: string}}
      responses:
        '302':
          description: Redirect to allowlisted Worthly client result page
        '400': {$ref: '#/components/responses/BadRequest'}
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://worthly.example.com/oauth2/authorize
          tokenUrl: https://worthly.example.com/oauth2/token
          refreshUrl: https://worthly.example.com/oauth2/token
          scopes:
            openid: OpenID identity
            profile: Basic owner profile
            worthly.read: Read Worthly financial data
            worthly.write: Modify Worthly metadata/settings
...
  responses:
...
    BadRequest:
      description: Bad request
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        correlationId: {type: string}